+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.18.0](https://github.com/perpetual-protocol/sdk-curie/compare/v1.17.0...v1.18.0) (2023-02-09)
+
+### Code Refactoring
+
+* refactor ClearingHouseConfig with twapInterval ([fb94711](https://github.com/perpetual-protocol/sdk-curie/commit/fb9471152edda4b2a6e9607e3fe4f13a9a111ff8))
+
+
+## [1.18.0-canary.0](https://github.com/perpetual-protocol/sdk-curie/compare/v1.17.0...v1.18.0-canary.0) (2023-02-09)
+
+
+### Code Refactoring
+
+* refactor ClearingHouseConfig with twapInterval ([fb94711](https://github.com/perpetual-protocol/sdk-curie/commit/fb9471152edda4b2a6e9607e3fe4f13a9a111ff8))
+
 ## [1.16.0](https://github.com/perpetual-protocol/sdk-curie/compare/v1.15.0...v1.16.0) (2022-10-31)